+++ src/gdpc/editor.py
@@ -4,17 +4,19 @@ Minecraft world through the GDMC HTTP interface."""
 
 from __future__ import annotations
 
-from typing import Dict, Sequence, Union, Optional, List, Iterable, Generator, Sized, cast
+from typing import Dict, Sequence, Union, Optional, List, Iterable, Generator, Sized, Any, cast
 from numbers import Integral
 from contextlib import contextmanager
 from copy import copy, deepcopy
 import random
 from concurrent import futures
 import logging
+import atexit
+import weakref
 
 import numpy as np
 import numpy.typing as npt
-from glm import ivec3
+from pyglm.glm import ivec3
 
 from .utils import eagerAll, OrderedByLookupDict
 from .vector_tools import Vec3iLike, Rect, Box, dropY
@@ -73,7 +75,7 @@ class Editor:
         self._multithreading = False
         self._multithreadingWorkers = multithreadingWorkers
         self.multithreading = multithreading # The property setter initializes the multithreading system.
-        self._bufferFlushFutures: List[futures.Future] = []
+        self._bufferFlushFutures: List[futures.Future[None]] = []
 
         self._doBlockUpdates = True
         self._spawnDrops     = False
@@ -83,17 +85,54 @@ class Editor:
         self._worldSlice: Optional[WorldSlice] = None
         self._worldSliceDecay: Optional[npt.NDArray[np.bool_]] = None
 
+        # We need to do some cleanup when the Editor gets deleted or when the program ends. In
+        # particular, we need to flush the block buffer and await any remaining buffer flush
+        # futures.
+        #
+        # The standard way to do this would be to put the cleanup code in __del__. But while this
+        # works for the editor-gets-deleted case, it does not work reliably for the program-ends
+        # case. __del__ does get called at program exit, but only when the interpreter is already
+        # half-shut-down. At that point, we can no longer send HTTP requests or schedule additional
+        # futures, so buffer flushing doesn't work anymore.
+        #
+        # One way to solve this would be to make Editor a context manager, but that would make it
+        # slightly harder to use (and it would be a breaking change). Instead, we solve it using an
+        # atexit handler. These get called early enough for buffer flushing to still work.
+        #
+        # We have to be careful about a few things:
+        # - We must use a weak reference to refer to the Editor. A normal reference would keep the
+        #   Editor alive and prevent the atexit handler from ever getting called.
+        # - We must create a new callable object for each Editor, so that when we unregister the
+        #   atexit handler, we only unregister the handler for the Editor that registered it. We
+        #   accomplish this using a lambda.
+        # - We need to store the lambda so that we can unregister it in __del__ and avoid a double
+        #   cleanup.
+
+        ref = weakref.ref(cast(Editor, self))
+        self._clean_up_handler = lambda: Editor._clean_up_at_exit(ref)
+        atexit.register(self._clean_up_handler)
+
+
+    @staticmethod
+    def _clean_up(editor: Editor) -> None:
+        """The real Editor destructor, called from both __del__ and the atexit handler."""
+        editor.flushBuffer()
+        editor.awaitBufferFlushes()
+
+
+    @staticmethod
+    def _clean_up_at_exit(editor_weakref: weakref.ref[Editor]) -> None:
+        editor = editor_weakref()
+        if editor is not None:
+            Editor._clean_up(editor)
+
 
     def __del__(self) -> None:
-        """Cleans up this Editor instance."""
-        # awaits any pending buffer flush futures and shuts down the buffer flush executor
-        self.multithreading = False
-        # Flush any remaining blocks in the buffer.
-        # This is purposefully done *after* disabling multithreading! This __del__ may be called at
-        # interpreter shutdown, and it appears that scheduling a new future at that point fails with
-        # "RuntimeError: cannot schedule new futures after shutdown" even if the executor has not
-        # actually shut down yet. For safety, the last buffer flush must be done on the main thread.
-        self.flushBuffer()
+        """Cleans up this Editor instance.\n
+        Flushes the block buffer and waits for any remaining buffer flush futures.
+        """
+        atexit.unregister(self._clean_up_handler)
+        Editor._clean_up(self)
 
 
     @property
@@ -266,7 +305,7 @@ class Editor:
         :meth:`.placeBlock`/:meth:`.placeBlockGlobal` will be given a block update, just like if
         they were placed by a player. We recommend leaving this on in most cases.
 
-        \*: The system is actually a bit more complex than this.
+        \\*: The system is actually a bit more complex than this.
         See https://minecraft.wiki/w/Block_update for the full details.
         """
         return self._doBlockUpdates
@@ -358,7 +397,7 @@ class Editor:
         return view
 
 
-    def runCommand(self, command: str, position: Optional[Vec3iLike]=None, syncWithBuffer=False) -> None:
+    def runCommand(self, command: str, position: Optional[Vec3iLike] = None, syncWithBuffer: bool = False) -> None:
         """Executes one or multiple Minecraft commands (separated by newlines).\n
         The leading "/" must be omitted.\n
         If buffering is enabled and ``syncWithBuffer`` is ``True``, the command is deferred until
@@ -377,7 +416,7 @@ class Editor:
         self.runCommandGlobal(command, position, syncWithBuffer)
 
 
-    def runCommandGlobal(self, command: str, position: Optional[Vec3iLike]=None, syncWithBuffer=False) -> None:
+    def runCommandGlobal(self, command: str, position: Optional[Vec3iLike] = None, syncWithBuffer: bool = False) -> None:
         """Executes one or multiple Minecraft commands (separated by newlines), ignoring :attr:`.transform`.\n
         The leading "/" must be omitted.\n
         If buffering is enabled and ``syncWithBuffer`` is ``True``, the command is deferred until
@@ -495,7 +534,7 @@ class Editor:
                 hasattr(position, "__len__")
                 and len(cast(Sized, position)) == 3
                 and hasattr(position, "__getitem__")
-                and isinstance(cast(Sequence, position)[0], Integral)
+                and isinstance(cast(Sequence[Any], position)[0], Integral)
             )
             else (self.transform * pos for pos in cast(Iterable[Vec3iLike], position))
         )
@@ -519,7 +558,7 @@ class Editor:
             hasattr(position, "__len__")
             and len(cast(Sized, position)) == 3
             and hasattr(position, "__getitem__")
-            and isinstance(cast(Sequence, position)[0], Integral)
+            and isinstance(cast(Sequence[Any], position)[0], Integral)
         ):
             return self._placeSingleBlockGlobal(ivec3(*cast(Vec3iLike, position)), block, replace)
 
@@ -646,7 +685,7 @@ class Editor:
         self._bufferFlushFutures = list(futures.wait(self._bufferFlushFutures, timeout).not_done)
 
 
-    def loadWorldSlice(self, rect: Optional[Rect]=None, heightmapTypes: Optional[Iterable[str]] = None, cache=False) -> WorldSlice:
+    def loadWorldSlice(self, rect: Optional[Rect]=None, heightmapTypes: Optional[Iterable[str]] = None, cache: bool = False) -> WorldSlice:
         """Loads the world slice for the given XZ-rectangle.\n
         The rectangle must be given in **global coordinates**; :attr:`.transform` is ignored.\n
         If ``rect`` is None, the world slice of the current build area is loaded.\n