+++ src/gdpc/interface.py
@@ -13,13 +13,15 @@ import logging
 import json
 import io
 
-from glm import ivec3
+import numpy as np
+import numpy.typing as npt
+from pyglm.glm import ivec3
 from nbt import nbt
 import requests
 from requests.exceptions import ConnectionError as RequestConnectionError
 
 from . import __url__
-from .utils import withRetries
+from .utils import withRetries, isIterable
 from .vector_tools import Vec2iLike, Vec3iLike, Box
 from .block import Block
 from . import exceptions
@@ -58,7 +60,16 @@ def _request(method: str, url: str, *args: Any, retries: int, **kwargs: Any) -> 
     return response
 
 
-def getBlocks(position: Vec3iLike, size: Optional[Vec3iLike] = None, dimension: Optional[str] = None, includeState=True, includeData=True, retries=0, timeout=None, host=DEFAULT_HOST) -> List[Tuple[ivec3, Block]]:
+def getBlocks(
+    position: Vec3iLike,
+    size: Optional[Vec3iLike] = None,
+    dimension: Optional[str] = None,
+    includeState: bool = True,
+    includeData: bool = True,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str =DEFAULT_HOST
+) -> List[Tuple[ivec3, Block]]:
     """Returns the blocks in the specified region.
 
     ``dimension`` can be one of {"overworld", "the_nether", "the_end"} (default "overworld").
@@ -84,7 +95,14 @@ def getBlocks(position: Vec3iLike, size: Optional[Vec3iLike] = None, dimension: 
     return [(ivec3(b["x"], b["y"], b["z"]), Block(b["id"], b.get("state", {}), b.get("data") if b.get("data") != "{}" else None)) for b in blockDicts]
 
 
-def getBiomes(position: Vec3iLike, size: Optional[Vec3iLike] = None, dimension: Optional[str] = None, retries=0, timeout=None, host=DEFAULT_HOST) -> List[Tuple[ivec3, str]]:
+def getBiomes(
+    position: Vec3iLike,
+    size: Optional[Vec3iLike] = None,
+    dimension: Optional[str] = None,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> List[Tuple[ivec3, str]]:
     """Returns the biomes in the specified region.
 
     ``dimension`` can be one of {"overworld", "the_nether", "the_end"} (default "overworld").
@@ -110,7 +128,16 @@ def getBiomes(position: Vec3iLike, size: Optional[Vec3iLike] = None, dimension: 
     return [(ivec3(b["x"], b["y"], b["z"]), str(b["id"])) for b in biomeDicts]
 
 
-def placeBlocks(blocks: Iterable[Tuple[Vec3iLike, Block]], dimension: Optional[str] = None, doBlockUpdates=True, spawnDrops=False, customFlags: str = "", retries=0, timeout=None, host=DEFAULT_HOST) -> List[Tuple[bool, Union[int, str]]]:
+def placeBlocks(
+    blocks: Iterable[Tuple[Vec3iLike, Block]],
+    dimension: Optional[str] = None,
+    doBlockUpdates: bool = True,
+    spawnDrops: bool = False,
+    customFlags: str = "",
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> List[Tuple[bool, Union[int, str]]]:
     """Places blocks in the world.
 
     Each element of ``blocks`` should be a tuple (position, block). Empty blocks (blocks without an
@@ -156,7 +183,13 @@ def placeBlocks(blocks: Iterable[Tuple[Vec3iLike, Block]], dimension: Optional[s
     return result
 
 
-def runCommand(command: str, dimension: Optional[str] = None, retries=0, timeout=None, host=DEFAULT_HOST) -> List[Tuple[bool, Optional[str]]]:
+def runCommand(
+    command: str,
+    dimension: Optional[str] = None,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> List[Tuple[bool, Optional[str]]]:
     """Executes one or multiple Minecraft commands (separated by newlines).
 
     The leading "/" must be omitted.
@@ -172,7 +205,11 @@ def runCommand(command: str, dimension: Optional[str] = None, retries=0, timeout
     return result
 
 
-def getBuildArea(retries=0, timeout=None, host=DEFAULT_HOST) -> Box:
+def getBuildArea(
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> Box:
     """Retrieves the build area that was specified with /setbuildarea in-game.
 
     Raises a :exc:`.BuildAreaNotSetError` if the build area was not specified yet.
@@ -202,7 +239,15 @@ def getBuildArea(retries=0, timeout=None, host=DEFAULT_HOST) -> Box:
     return Box.between(fromPoint, toPoint)
 
 
-def getChunks(position: Vec2iLike, size: Optional[Vec2iLike] = None, dimension: Optional[str] = None, asBytes=False, retries=0, timeout=None, host=DEFAULT_HOST) -> Union[str, bytes]:
+def getChunks(
+    position: Vec2iLike,
+    size: Optional[Vec2iLike] = None,
+    dimension: Optional[str] = None,
+    asBytes: bool = False,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> Union[str, bytes]:
     """Returns raw chunk data.
 
     ``position`` specifies the position in chunk coordinates, and ``size`` specifies how many chunks
@@ -229,7 +274,21 @@ def getChunks(position: Vec2iLike, size: Optional[Vec2iLike] = None, dimension: 
     return response.content if asBytes else response.text
 
 
-def placeStructure(structureData: Union[bytes, nbt.NBTFile], position: Vec3iLike, mirror: Optional[Vec2iLike] = None, rotate: Optional[int] = None, pivot: Optional[Vec3iLike] = None, includeEntities: Optional[bool] = None, dimension: Optional[str] = None, doBlockUpdates=True, spawnDrops=False, customFlags: str = "", retries=0, timeout=None, host=DEFAULT_HOST) -> None:
+def placeStructure(
+    structureData: Union[bytes, nbt.NBTFile],
+    position: Vec3iLike,
+    mirror: Optional[Vec2iLike] = None,
+    rotate: Optional[int] = None,
+    pivot: Optional[Vec3iLike] = None,
+    includeEntities: Optional[bool] = None,
+    dimension: Optional[str] = None,
+    doBlockUpdates: bool = True,
+    spawnDrops: bool = False,
+    customFlags: str = "",
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> None:
     """Places a structure defined using the Minecraft structure format in the world.
 
     ``structureData`` should be a string of bytes in the Minecraft structure file format, the format used by the
@@ -285,7 +344,16 @@ def placeStructure(structureData: Union[bytes, nbt.NBTFile], position: Vec3iLike
     return response.json()
 
 
-def getStructure(position: Vec3iLike, size: Vec3iLike, dimension: Optional[str] = None, includeEntities: Optional[bool] = None, returnCompressed: Optional[bool] = True, retries=0, timeout=None, host=DEFAULT_HOST) -> bytes:
+def getStructure(
+    position: Vec3iLike,
+    size: Vec3iLike,
+    dimension: Optional[str] = None,
+    includeEntities: Optional[bool] = None,
+    returnCompressed: Optional[bool] = True,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> bytes:
     """Returns the specified area in the Minecraft structure file format (an NBT byte string).
 
     The Minecraft structure file format is the format used by the in-game structure blocks. Structures in this format
@@ -316,7 +384,85 @@ def getStructure(position: Vec3iLike, size: Vec3iLike, dimension: Optional[str] 
     return response.content
 
 
-def getEntities(selector: Optional[str] = None, includeData: bool = True, dimension: Optional[str] = None, retries=0, timeout=None, host=DEFAULT_HOST) -> Any:
+def getHeightmap(
+    position: Optional[Vec3iLike] = None,
+    size: Optional[Vec3iLike] = None,
+    heightmapType: Optional[str] = None,
+    blocks: Optional[Iterable[str]] = None,
+    yMin: Optional[int] = None,
+    yMax: Optional[int] = None,
+    dimension: Optional[str] = None,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> npt.NDArray[np.int_]:
+    """
+    Returns heightmap of the given type within the current build area.
+
+    This endpoint supports four of `Minecraft's built-in heightmap types <https://minecraft.wiki/w/Heightmap>`_:
+
+    * ``'WORLD_SURFACE'``: Height of the surface ignoring air blocks.
+    * ``'OCEAN_FLOOR'``: Height of the surface ignoring air, water, and lava.
+    * ``'MOTION_BLOCKING'``: Height of the surface ignoring blocks that don't have movement collision (air, flowers,
+      ferns, etc.), except for water and lava.
+    * ``'MOTION_BLOCKING_NO_LEAVES'``: Same as ``'MOTION_BLOCKING'``, but also ignores
+      `leaves <https://minecraft.wiki/w/Leaves>`_.
+
+    Additionally, the GDMC-HTTP mod provides two extra heightmap types,
+    which can only be retrieved using this endpoint:
+
+    * ``'MOTION_BLOCKING_NO_PLANTS'``: Same as ``'MOTION_BLOCKING_NO_LEAVES'``, but also excludes various biological
+      block types. For a full list, refer to the `GDMC-HTTP documentation
+      <https://github.com/Niels-NTG/gdmc_http_interface/blob/master/docs/Endpoints.md#heightmap-preset-types>`_.
+    * ``'OCEAN_FLOOR_NO_PLANTS'``: Same as ``'OCEAN_FLOOR'``, except it also excludes everything that is part of
+      ``'MOTION_BLOCKING_NO_PLANTS'``.
+
+    Instead of a heightmap type, you can also submit a list of block IDs using the ``blocks`` parameter.
+    These are the blocks that should be considered "transparent" when the heightmap is calculated.
+    **Note:** Air blocks (``'minecraft:air'``, ``'minecraft:cave_air'``) aren't included by default.
+
+    Using the ``yMin`` and ``yMax`` parameters, the lower and/or upper limit of the heightmap calculation can be
+    constrained. This can be useful for creating heightmaps of overworld caves or the Nether dimension.
+    **Only works if used in conjunction with the** ``blocks`` **parameter.**
+    """
+    customBlocksQuery = ""
+    yBoundsQuery = ""
+    if isIterable(blocks):
+        customBlocksQuery = ",".join(cast(Iterable[str], blocks))
+        yBoundsQuery = f'{yMin if isinstance(yMin, int) else ""}..{yMax if isinstance(yMax, int) else ""}'
+    elif heightmapType is None:
+        # Default to heightmap type WORLD_SURFACE if both `heightMapType` and `blocks` parameters aren't set.
+        heightmapType = 'WORLD_SURFACE'
+    parameters: Dict[str, Any] = {
+        'type': heightmapType,
+        'dimension': dimension,
+        'blocks': customBlocksQuery,
+        'yBounds': yBoundsQuery,
+    }
+    if position is not None:
+        parameters.update({
+            'x': position[0],
+            'y': position[1],
+            'z': position[2],
+        })
+    if size is not None:
+        parameters.update({
+            'dx': size[0],
+            'dy': size[1],
+            'dz': size[2],
+        })
+    response = _request(method='GET', url=f'{host}/heightmap', params=parameters, retries=retries, timeout=timeout)
+    return np.asarray(response.json(), dtype=np.int_)
+
+
+def getEntities(
+    selector: Optional[str] = None,
+    includeData: bool = True,
+    dimension: Optional[str] = None,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> Any:
     url = f'{host}/entities'
     parameters = {
         'selector': selector,
@@ -327,7 +473,14 @@ def getEntities(selector: Optional[str] = None, includeData: bool = True, dimens
     return response.json()
 
 
-def getPlayers(selector: Optional[str] = None, includeData: bool = True, dimension: Optional[str] = None, retries=0, timeout=None, host=DEFAULT_HOST) -> Any:
+def getPlayers(
+    selector: Optional[str] = None,
+    includeData: bool = True,
+    dimension: Optional[str] = None,
+    retries: int = 0,
+    timeout: Any = None,
+    host: str = DEFAULT_HOST
+) -> Any:
     url = f'{host}/players'
     parameters = {
         'selector': selector,
@@ -338,6 +491,6 @@ def getPlayers(selector: Optional[str] = None, includeData: bool = True, dimensi
     return response.json()
 
 
-def getVersion(retries=0, timeout=None, host=DEFAULT_HOST) -> str:
+def getVersion(retries: int = 0, timeout: Any =None, host: str = DEFAULT_HOST) -> str:
     """Returns the Minecraft version as a string."""
     return _request("GET", f"{host}/version", retries=retries, timeout=timeout).text